chore(wallet)_: community deployment related types moved to wallet requests package
`DeploymentParameters` and `DeploymentDetails` types moved from `communitytokens` package
to `requests` of the wallet service.
